Move make_builder into mod.rs
